+++ src/platforms/avr/int.h
@@ -1,8 +1,5 @@
 #pragma once
 
-#include <stdint.h>
-#include <stddef.h>
-
 namespace fl {
     // On AVR: int is 16-bit, long is 32-bit — match stdint sizes manually
     typedef int i16;
@@ -11,7 +8,7 @@ namespace fl {
     typedef unsigned long u32;
     typedef long long i64;
     typedef unsigned long long u64;
-    // Use standard types directly to ensure exact compatibility
-    typedef size_t size;
-    typedef uintptr_t uptr;
+    // AVR is 8-bit: pointers are 16-bit
+    typedef unsigned int size;   // size_t equivalent (16-bit on AVR)
+    typedef unsigned int uptr;   // uintptr_t equivalent (16-bit on AVR)
 }
